requirements.txt
streamlit
scikit-learn
joblib
datasets
transformers
torch
evaluate
pandas
numpy
lime
shap
pillow